Extracted logic to new trait
