+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/ModuleResolvers/HasMarkdownDocumentationModuleResolverTrait.php
@@ -5,11 +5,12 @@ declare(strict_types=1);
 namespace GraphQLAPI\GraphQLAPI\ModuleResolvers;
 
 use GraphQLAPI\GraphQLAPI\ContentProcessors\ContentParserOptions;
-use InvalidArgumentException;
-use GraphQLAPI\GraphQLAPI\Facades\ContentProcessors\MarkdownContentParserFacade;
+use GraphQLAPI\GraphQLAPI\ContentProcessors\MarkdownContentRetrieverTrait;
 
 trait HasMarkdownDocumentationModuleResolverTrait
 {
+    use MarkdownContentRetrieverTrait;
+
     /**
      * The module slug
      */
@@ -38,35 +39,18 @@ trait HasMarkdownDocumentationModuleResolverTrait
     public function getDocumentation(string $module): ?string
     {
         if ($markdownFilename = $this->getMarkdownFilename($module)) {
-            $markdownContentParser = MarkdownContentParserFacade::getInstance();
-            // Inject the place to look for the documentation
-            $markdownContentParser->setBaseDir($this->getBaseDir());
-            $markdownContentParser->setBaseURL($this->getBaseURL());
-            try {
-                return $markdownContentParser->getContent(
-                    'modules/' . $markdownFilename,
-                    'modules',
-                    [
-                        ContentParserOptions::TAB_CONTENT => true,
-                    ]
-                );
-            } catch (InvalidArgumentException) {
-                return sprintf(
+            return $this->getMarkdownContent(
+                'modules/' . $markdownFilename,
+                'modules',
+                [
+                    ContentParserOptions::TAB_CONTENT => true,
+                ],
+                sprintf(
                     '<p>%s</p>',
                     \__('Oops, the documentation for this module is not available', 'graphql-api')
-                );
-            }
+                )
+            );
         }
         return null;
     }
-
-    /**
-     * Get the dir where to look for the documentation.
-     */
-    abstract protected function getBaseDir(): string;
-
-    /**
-     * Get the URL where to look for the documentation.
-     */
-    abstract protected function getBaseURL(): string;
 }
